authentication-module/requirements/development.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
Faker==18.10.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
pika==1.3.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pyotp==2.8.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
